fixed two bugs which very very hard to track:
FIX crash when >200 nodes + > 500 connections when CTRL+A + 'x' <- remove items
FIX bug when having two GraphicsViews open and a segfault pops up - if node after node is deleted - after a
while.

the solution was:
 - use valgrind with a debug enabled qt 4.4
 - find out what caused the error using valgrind while manually triggering the bug mentioned above

next was to look at the valgrind trace which pointed out that a already removed (not deleted) item
was triggered to be painted -> QGraphicsItem::paint() call which internally used scene()->somememeber
to find out about some global scene settings. since the item was in no scene anymore this scene() lookup
of course produced a segmentation fault. very odd since the item was not in any scene i didn't expect
Qt to draw it this is why i didn't put any if (scene() == NULL) return; code into it.

is this a bug in qt4.4? it seems to be triggered very very rarely. say when deleting 40-400 items in one
go i could trigger it frequently but only with a few connection deletes nearly never...?!

--

debug code will be removed soon and functionality will be resotred on with the next commit